MINT-1788 Move redux/modals to core
redux/modals is a pure state part. It is needed by some core actions
and can't be removed from core totally.
Modals reducer can be replaced in different environment.
