+++ src/components/Deposits/Deposit/Deposit.jsx
@@ -13,7 +13,7 @@ import { getDeposit } from '@chronobank/core/redux/mainWallet/selectors'
 import Button from 'components/common/ui/Button/Button'
 import IPFSImage from 'components/common/IPFSImage/IPFSImage'
 import TokenValue from 'components/common/TokenValue/TokenValue'
-import { modalsOpen } from 'redux/modals/actions'
+import { modalsOpen } from '@chronobank/core/redux/modals/actions'
 import { DUCK_TOKENS } from '@chronobank/core/redux/tokens/constants'
 import TokenModel from '@chronobank/core/models/tokens/TokenModel'
 import { TOKEN_ICONS } from 'assets'
