+++ src/components/Deposits/DepositWarningWidget/DepositWarningWidget.jsx
@@ -13,7 +13,7 @@ import { TIME } from '@chronobank/core/dao/constants'
 import { DUCK_ASSETS_HOLDER } from '@chronobank/core/redux/assetsHolder/constants'
 import AllowanceModel from '@chronobank/core/models/wallet/AllowanceModel'
 import { getDeposit } from '@chronobank/core/redux/mainWallet/selectors'
-import { modalsOpen } from 'redux/modals/actions'
+import { modalsOpen } from '@chronobank/core/redux/modals/actions'
 import WalletModel from '@chronobank/core/models/wallet/WalletModel'
 import { getMainEthWallet } from '@chronobank/core/redux/wallets/selectors/models'
 import PropTypes from 'prop-types'
