+++ src/components/assetsManager/AddTokenDialog/AddTokenDialog.jsx
@@ -8,7 +8,7 @@ import { Translate } from 'react-redux-i18n'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import ModalDialog from 'components/dialogs/ModalDialog'
-import { modalsClose } from 'redux/modals/actions'
+import { modalsClose } from '@chronobank/core/redux/modals/actions'
 import AddTokenForm, { prefix } from './AddTokenForm'
 
 function mapDispatchToProps (dispatch) {
