+++ src/components/assetsManager/AddTokenDialog/AddTokenForm.jsx
@@ -25,7 +25,7 @@ import FileModel from '@chronobank/core/models/FileSelect/FileModel'
 import TokenModel from '@chronobank/core/models/tokens/TokenModel'
 import { createAsset } from '@chronobank/core/redux/assetsManager/actions'
 import { DUCK_ASSETS_MANAGER } from '@chronobank/core/redux/assetsManager/constants'
-import { modalsOpen } from 'redux/modals/actions'
+import { modalsOpen } from '@chronobank/core/redux/modals/actions'
 import IPFSImage from 'components/common/IPFSImage/IPFSImage'
 import ipfs from '@chronobank/core/utils/IPFS'
 import FeeModel from '@chronobank/core/models/tokens/FeeModel'
