+++ src/components/assetsManager/AssetManager/AssetManager.jsx
@@ -11,7 +11,7 @@ import { Translate } from 'react-redux-i18n'
 import { connect } from 'react-redux'
 import { DUCK_ASSETS_MANAGER } from '@chronobank/core/redux/assetsManager/constants'
 import { createPlatform } from '@chronobank/core/redux/assetsManager/actions'
-import { modalsOpen } from 'redux/modals/actions'
+import { modalsOpen } from '@chronobank/core/redux/modals/actions'
 import HistoryTable from 'components/assetsManager/HistoryTable/HistoryTable'
 import PlatformInfo from 'components/assetsManager/PlatformInfo/PlatformInfo'
 import PlatformsList from 'components/assetsManager/PlatformsList/PlatformsList'
