+++ src/components/assetsManager/BlacklistDialog/BlacklistDialog.jsx
@@ -7,7 +7,7 @@ import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import ModalDialog from 'components/dialogs/ModalDialog'
-import { modalsClose } from 'redux/modals/actions'
+import { modalsClose } from '@chronobank/core/redux/modals/actions'
 import { Translate } from 'react-redux-i18n'
 import BlacklistForm from 'components/assetsManager/BlacklistForm/BlacklistForm'
 import { restrictUser, unrestrictUser } from '@chronobank/core/redux/assetsManager/actions'
