+++ src/components/assetsManager/BlockAssetDialog/BlockAssetDialog.jsx
@@ -7,7 +7,7 @@ import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import ModalDialog from 'components/dialogs/ModalDialog'
-import { modalsClose } from 'redux/modals/actions'
+import { modalsClose } from '@chronobank/core/redux/modals/actions'
 import { Translate } from 'react-redux-i18n'
 import { changePauseStatus } from '@chronobank/core/redux/assetsManager/actions'
 import TokenModel from '@chronobank/core/models/tokens/TokenModel'
