+++ src/components/assetsManager/PlatformInfo/PlatformInfo.jsx
@@ -15,7 +15,7 @@ import { connect } from 'react-redux'
 import { Translate } from 'react-redux-i18n'
 import { getFee, getManagersForAssetSymbol } from '@chronobank/core/redux/assetsManager/actions'
 import { DUCK_ASSETS_MANAGER } from '@chronobank/core/redux/assetsManager/constants'
-import { modalsOpen } from 'redux/modals/actions'
+import { modalsOpen } from '@chronobank/core/redux/modals/actions'
 import ReissueAssetForm from 'components/assetsManager/ReissueAssetForm/ReissueAssetForm'
 import { getSelectedToken } from '@chronobank/core/redux/assetsManager/selectors'
 import TokenModel from '@chronobank/core/models/tokens/TokenModel'
