+++ src/components/assetsManager/RevokeDialog/RevokeDialog.jsx
@@ -8,7 +8,7 @@ import { Translate } from 'react-redux-i18n'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import ModalDialog from 'components/dialogs/ModalDialog'
-import { modalsClose } from 'redux/modals/actions'
+import { modalsClose } from '@chronobank/core/redux/modals/actions'
 import RevokeForm, { prefix } from './RevokeForm'
 
 function mapDispatchToProps (dispatch) {
