+++ src/components/common/ModalStack/ModalStack.spec.js
@@ -7,7 +7,7 @@ import React from 'react'
 import Immutable from 'immutable'
 import { Provider } from 'react-redux'
 import { mockStore } from 'specsInit'
-import { DUCK_MODALS } from 'redux/modals/constants'
+import { DUCK_MODALS } from '@chronobank/redux/modals/constants'
 import { mount } from 'enzyme'
 import ModalDialog from 'components/dialogs/ModalDialog'
 import ModalStack from 'components/common/ModalStack/ModalStack'
