+++ src/components/dashboard/DepositTokens/DepositTokensForm.jsx
@@ -11,7 +11,7 @@ import { isTestingNetwork } from '@chronobank/login/network/settings'
 import { DUCK_NETWORK } from '@chronobank/login/redux/network/constants'
 import web3Converter from '@chronobank/core/utils/Web3Converter'
 import { TOKEN_ICONS } from 'assets'
-import { modalsOpen } from 'redux/modals/actions'
+import { modalsOpen } from '@chronobank/core/redux/modals/actions'
 import BigNumber from 'bignumber.js'
 import Preloader from 'components/common/Preloader/Preloader'
 import TokenValue from 'components/common/TokenValue/TokenValue'
