+++ src/components/dashboard/DepositTokens/DepositTokensModal.jsx
@@ -9,7 +9,7 @@ import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
 import { change } from 'redux-form'
 import { depositAsset, initAssetsHolder, withdrawAsset } from '@chronobank/core/redux/assetsHolder/actions'
-import { modalsClose } from 'redux/modals/actions'
+import { modalsClose } from '@chronobank/core/redux/modals/actions'
 import { mainApprove } from '@chronobank/core/redux/wallets/actions'
 import ModalDialog from 'components/dialogs/ModalDialog'
 import {
