+++ src/components/dialogs/AddTokenDialog/AddTokenDialog.jsx
@@ -14,7 +14,7 @@ import { ACCEPT_IMAGES } from '@chronobank/core/models/FileSelect/FileExtension'
 import TokenModel from '@chronobank/core/models/tokens/TokenModel'
 import { addToken, formTokenLoadMetaData } from '@chronobank/core/redux/settings/erc20/tokens/actions'
 import { DUCK_SESSION } from '@chronobank/core/redux/session/constants'
-import { modalsClose } from 'redux/modals/actions'
+import { modalsClose } from '@chronobank/core/redux/modals/actions'
 import FileSelect from 'components/common/FileSelect/FileSelect'
 import IPFSImage from 'components/common/IPFSImage/IPFSImage'
 import { DUCK_TOKENS } from '@chronobank/core/redux/tokens/constants'
