+++ src/components/dialogs/ConfirmTxDialog/ConfirmTxDialog.jsx
@@ -11,7 +11,7 @@ import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
 import { Translate } from 'react-redux-i18n'
-import { modalsClear, modalsClose } from 'redux/modals/actions'
+import { modalsClear, modalsClose } from '@chronobank/core/redux/modals/actions'
 import { ETH } from '@chronobank/core/dao/constants'
 import TxEntryModel from '@chronobank/core/models/TxEntryModel'
 import { LogTxModel } from '@chronobank/core/models'
