+++ src/components/dialogs/ModalDialogBase/ModalDialogBase.js
@@ -8,7 +8,7 @@ import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import { Translate } from 'react-redux-i18n'
 import { connect } from 'react-redux'
-import { modalsClose } from 'redux/modals/actions'
+import { modalsClose } from '@chronobank/core/redux/modals/actions'
 import ModalDialog from '../ModalDialog'
 
 import './ModalDialogBase.scss'
