+++ src/components/dialogs/PublishPollDialog/PublishPollDialog.jsx
@@ -8,7 +8,7 @@ import Button from 'components/common/ui/Button/Button'
 import React, { PureComponent } from 'react'
 import { Translate } from 'react-redux-i18n'
 import { connect } from 'react-redux'
-import { modalsClose } from 'redux/modals/actions'
+import { modalsClose } from '@chronobank/core/redux/modals/actions'
 import { PTPoll } from '@chronobank/core/redux/voting/types'
 import { activatePoll } from '@chronobank/core/redux/voting/thunks'
 import ModalDialog from '../ModalDialog'
