+++ src/components/dialogs/UserActiveDialog/UserActiveDialog.jsx
@@ -10,7 +10,7 @@ import { Translate } from 'react-redux-i18n'
 import userMonitorService from 'user/monitorService'
 import { connect } from 'react-redux'
 import { logoutAndNavigateToRoot } from 'redux/ui/thunks'
-import { modalsClose } from 'redux/modals/actions'
+import { modalsClose } from '@chronobank/core/redux/modals/actions'
 import { stopUserMonitor } from 'redux/ui/actions'
 import ModalDialog from 'components/dialogs/ModalDialog'
 import Timer from 'components/common/Timer/Timer'
