+++ src/components/dialogs/wallet/EditOwnersDialog/EditOwnersDialog.js
@@ -7,7 +7,7 @@ import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
 import { addOwner, removeOwner } from '@chronobank/core/redux/multisigWallet/actions'
-import { modalsClose } from 'redux/modals/actions'
+import { modalsClose } from '@chronobank/core/redux/modals/actions'
 import EditManagersBaseForm from 'components/forms/EditManagersBaseForm/EditManagersBaseForm'
 import ModalDialog from 'components/dialogs/ModalDialog'
 import MultisigEthWalletModel from '@chronobank/core/models/wallet/MultisigEthWalletModel'
