+++ src/components/voting/PollEditForm/PollEditForm.jsx
@@ -27,7 +27,7 @@ import { TX_CREATE_POLL } from '@chronobank/core/dao/constants/VotingManagerDAO'
 import { DUCK_SESSION } from '@chronobank/core/redux/session/constants'
 import { createPoll } from '@chronobank/core/redux/voting/thunks'
 import { DUCK_VOTING } from '@chronobank/core/redux/voting/constants'
-import { modalsClose } from 'redux/modals/actions'
+import { modalsClose } from '@chronobank/core/redux/modals/actions'
 import Amount from '@chronobank/core/models/Amount'
 import { FEE_RATE_MULTIPLIER } from '@chronobank/core/redux/mainWallet/constants'
 import { TIME } from '@chronobank/core/dao/constants'
