+++ src/components/wallet/AddWalletWidget/WalletSettingsForm/WalletSettingsForm.jsx
@@ -12,7 +12,7 @@ import { Translate } from 'react-redux-i18n'
 import { Field, formPropTypes, reduxForm } from 'redux-form/immutable'
 import ModalDialog from 'components/dialogs/ModalDialog'
 import { setMultisigWalletName } from '@chronobank/core/redux/multisigWallet/actions'
-import { modalsClose } from 'redux/modals/actions'
+import { modalsClose } from '@chronobank/core/redux/modals/actions'
 import { PTWallet } from '@chronobank/core/redux/wallet/types'
 import { setWalletName } from '@chronobank/core/redux/wallets/actions'
 import { FORM_WALLET_SETTINGS } from 'components/constants'
