+++ src/components/wallet/WalletWidget/WalletWidget.jsx
@@ -9,7 +9,7 @@ import { Link } from 'react-router'
 import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
 import { selectWallet } from '@chronobank/core/redux/wallet/actions'
-import { modalsOpen } from 'redux/modals/actions'
+import { modalsOpen } from '@chronobank/core/redux/modals/actions'
 import { Translate } from 'react-redux-i18n'
 import { TOKEN_ICONS } from 'assets'
 import { DUCK_TOKENS } from '@chronobank/core/redux/tokens/constants'
