+++ src/components/wallet/WalletWidgetDetail/WalletWidgetDetail.jsx
@@ -7,7 +7,7 @@ import PropTypes from 'prop-types'
 import TokenModel from '@chronobank/core/models/tokens/TokenModel'
 import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
-import { modalsOpen } from 'redux/modals/actions'
+import { modalsOpen } from '@chronobank/core/redux/modals/actions'
 import { Translate } from 'react-redux-i18n'
 import { TOKEN_ICONS } from 'assets'
 import Button from 'components/common/ui/Button/Button'
