+++ src/layouts/Markup.jsx
@@ -17,7 +17,7 @@ import { closeNotifier } from '@chronobank/core/redux/notifier/actions'
 import { DUCK_NOTIFIER } from '@chronobank/core/redux/notifier/constants'
 import theme from 'styles/themes/default'
 import { DUCK_SIDES } from 'redux/sides/constants'
-import { DUCK_MODALS } from 'redux/modals/constants'
+import { DUCK_MODALS } from '@chronobank/core/redux/modals/constants'
 import MuiThemeProvider from '@material-ui/core/es/styles/MuiThemeProvider'
 import IconButton from '@material-ui/core/es/IconButton/IconButton'
 import DrawerMainMenu from 'layouts/partials/DrawerMainMenu/DrawerMainMenu'
