+++ src/layouts/partials/DepositsContent/DepositsContent.jsx
@@ -13,7 +13,7 @@ import { getDeposit } from '@chronobank/core/redux/mainWallet/selectors'
 import Amount from '@chronobank/core/models/Amount'
 import DepositsList from 'components/Deposits/DepositsList/DepositsList'
 import Button from 'components/common/ui/Button/Button'
-import { modalsOpen } from 'redux/modals/actions'
+import { modalsOpen } from '@chronobank/core/redux/modals/actions'
 import DepositWarningWidget from 'components/Deposits/DepositWarningWidget/DepositWarningWidget'
 import { prefix } from './lang'
 import './DepositsContent.scss'
