+++ src/layouts/partials/DrawerMainMenu/DrawerMainMenu.jsx
@@ -22,7 +22,7 @@ import chronoWalletLogoSVG from 'assets/img/chronowallettext-white.svg'
 import ProfileImage from 'components/common/ProfileImage/ProfileImage'
 import exitSvg from 'assets/img/exit-white.svg'
 import { sidesCloseAll, sidesOpen } from 'redux/sides/actions'
-import { modalsOpen } from 'redux/modals/actions'
+import { modalsOpen } from '@chronobank/core/redux/modals/actions'
 import { getAccountAvatar, getAccountName } from '@chronobank/core/redux/persistAccount/utils'
 import { MENU_TOKEN_MORE_INFO_PANEL_KEY } from 'redux/sides/constants'
 import { getWalletsLength } from '@chronobank/core/redux/wallets/selectors/wallets'
