+++ src/layouts/partials/ProfileContent/ProfileContent.jsx
@@ -12,7 +12,7 @@ import ProfileModel from '@chronobank/core/models/ProfileModel'
 import React, { PureComponent } from 'react'
 import { logoutAndNavigateToRoot } from 'redux/ui/thunks'
 import { getBlockchainAddressesList } from '@chronobank/core/redux/session/selectors'
-import { modalsOpen } from 'redux/modals/actions'
+import { modalsOpen } from '@chronobank/core/redux/modals/actions'
 import IPFSImage from 'components/common/IPFSImage/IPFSImage'
 import CopyIcon from 'components/dashboard/MicroIcon/CopyIcon'
 import QRIcon from 'components/dashboard/MicroIcon/QRIcon'
